LayoutTests/crypto/subtle/ecdsa-import-compressed-spki-key-p384-expected.txt

Test importing a P-384 SPKI ECDSA COMPRESSED key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS publicKey.toString() is '[object CryptoKey]'
PASS publicKey.type is 'public'
PASS publicKey.extractable is true
PASS publicKey.algorithm.name is 'ECDSA'
PASS publicKey.algorithm.namedCurve is 'P-384'
PASS publicKey.usages is [ ]
PASS successfullyParsed is true

TEST COMPLETE
